refactor: rename client
