+++ packages/core/src/awsService/apprunner/explorer/apprunnerNode.ts
@@ -8,7 +8,7 @@ import { AWSTreeNodeBase } from '../../../shared/treeview/nodes/awsTreeNodeBase'
 import { AppRunnerServiceNode } from './apprunnerServiceNode'
 import { PlaceholderNode } from '../../../shared/treeview/nodes/placeholderNode'
 import * as nls from 'vscode-nls'
-import { AppRunnerClient } from '../../../shared/clients/apprunnerClient'
+import { AppRunnerClient } from '../../../shared/clients/apprunner'
 import { getPaginatedAwsCallIter } from '../../../shared/utilities/collectionUtils'
 import { AppRunner } from 'aws-sdk'
 import { PollingSet } from '../../../shared/utilities/pollingSet'
