+++ packages/core/src/awsService/apprunner/wizards/codeRepositoryWizard.ts
@@ -11,7 +11,7 @@ import { GitExtension } from '../../../shared/extensions/git'
 import * as vscode from 'vscode'
 import { WizardForm } from '../../../shared/wizards/wizardForm'
 import { createVariablesPrompter } from '../../../shared/ui/common/variablesPrompter'
-import { AppRunnerClient } from '../../../shared/clients/apprunnerClient'
+import { AppRunnerClient } from '../../../shared/clients/apprunner'
 import { makeDeploymentButton } from './deploymentButton'
 import { createLabelQuickPick, createQuickPick, QuickPickPrompter } from '../../../shared/ui/pickerPrompter'
 import { createInputBox, InputBoxPrompter } from '../../../shared/ui/inputPrompter'
