+++ packages/core/src/test/awsService/apprunner/explorer/apprunnerNode.test.ts
@@ -9,7 +9,7 @@ import * as sinon from 'sinon'
 import { AppRunner } from 'aws-sdk'
 import { AppRunnerNode } from '../../../../awsService/apprunner/explorer/apprunnerNode'
 import { AppRunnerServiceNode } from '../../../../awsService/apprunner/explorer/apprunnerServiceNode'
-import { AppRunnerClient } from '../../../../shared/clients/apprunnerClient'
+import { AppRunnerClient } from '../../../../shared/clients/apprunner'
 import { PlaceholderNode } from '../../../../shared/treeview/nodes/placeholderNode'
 import { AWSTreeNodeBase } from '../../../../shared/treeview/nodes/awsTreeNodeBase'
 import { installFakeClock } from '../../../testUtil'
