+++ packages/core/src/test/awsService/apprunner/explorer/apprunnerServiceNode.test.ts
@@ -8,7 +8,7 @@ import * as sinon from 'sinon'
 import { AppRunner } from 'aws-sdk'
 import { AppRunnerNode } from '../../../../awsService/apprunner/explorer/apprunnerNode'
 import { AppRunnerServiceNode } from '../../../../awsService/apprunner/explorer/apprunnerServiceNode'
-import { DefaultAppRunnerClient } from '../../../../shared/clients/apprunnerClient'
+import { AppRunnerClient } from '../../../../shared/clients/apprunner'
 import { DefaultCloudWatchLogsClient } from '../../../../shared/clients/cloudWatchLogsClient'
 import { asyncGenerator } from '../../../../shared/utilities/collectionUtils'
 import { AWSTreeNodeBase } from '../../../../shared/treeview/nodes/awsTreeNodeBase'
@@ -16,7 +16,7 @@ import { stub } from '../../../utilities/stubber'
 import { getLabel } from '../../../../shared/treeview/utils'
 
 describe('AppRunnerServiceNode', function () {
-    let mockApprunnerClient: ReturnType<typeof stub<DefaultAppRunnerClient>>
+    let mockApprunnerClient: ReturnType<typeof stub<AppRunnerClient>>
     let mockParentNode: AppRunnerNode
     let node: AppRunnerServiceNode
 
@@ -36,7 +36,7 @@ describe('AppRunnerServiceNode', function () {
         const cloudwatchClient = stub(DefaultCloudWatchLogsClient, { regionCode: 'us-east-1' })
         cloudwatchClient.describeLogGroups.returns(asyncGenerator([{ logGroupName: 'logs' }]))
 
-        mockApprunnerClient = stub(DefaultAppRunnerClient, { regionCode: 'us-east-1' })
+        mockApprunnerClient = stub(AppRunnerClient, { regionCode: 'us-east-1' })
         mockApprunnerClient.listOperations.resolves({ OperationSummaryList: [] })
         mockParentNode = stub(AppRunnerNode, {
             regionCode: '',
